🚚 rename: Community folder classification
